toma los productos
